🎨 Palette: Fix Python syntax and import errors
- Fixed an undefined name error in `src/math_kernel/service.py` where `f64` was used instead of `float`.
- Fixed a class naming convention in `src/shared/grpc_manager.py` (`gRPCManager` -> `GRPCManager`).
- Fixed `alembic/env.py` and `src/api/dependencies.py` to follow the standard top-level import conventions.

+++ src/math_kernel/service.py
@@ -24,7 +24,7 @@ except ImportError:
 
 class BlackScholesSolver:
     """Analytical Black-Scholes solver using the Rust-core."""
-    def compute(self, s: f64, k: f64, t: f64, r: f64, sigma: f64, is_call: bool = True) -> float:
+    def compute(self, s: float, k: float, t: float, r: float, sigma: float, is_call: bool = True) -> float:
         if bsopt_core:
             if is_call:
                 return bsopt_core.bs_call(s, k, t, r, sigma)
